requirements.txt
Flask==1.1.2
google-cloud-datastore==1.7.3
json2html==1.3.0
requests_oauthlib==1.3.0